Move distributions to key_generator.hpp
