+++ benchmarks/hash_table/static_multimap/count_bench.cu
@@ -15,7 +15,6 @@
  */
 
 #include <defaults.hpp>
-#include <distribution.hpp>
 #include <key_generator.hpp>
 #include <utils.hpp>
